extract CelGoogleMapsViewer.mjs and add verison
